+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Unreleased
 
+* Generate a suppressable warning when a component extends a class that is also annotated with `@ArezComponent`.
+
 ### [v0.201](https://github.com/arez/arez/tree/v0.201) (2021-11-10) · [Full Changelog](https://github.com/spritz/spritz/compare/v0.200...v0.201)
 
 Changes in this release: